auth

+++ swa-projektarbeit/src/main/java/de/hsos/swa/studiom/StudyGroupManagement/boundary/rest/Group/GroupGroupIDRessource.java
@@ -2,15 +2,17 @@
  * @author Joana Wegener
  * @email joana.wegener@hs-osnabrueck.de
  * @create date 2022-01-22 20:09:50
- * @modify date 2022-02-01 11:28:28
+ * @modify date 2022-02-01 16:22:29
  * @desc [description]
  */
 package de.hsos.swa.studiom.StudyGroupManagement.boundary.rest.Group;
 
 import java.util.Optional;
 
+import javax.annotation.security.RolesAllowed;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
+
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
@@ -21,10 +23,13 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import javax.ws.rs.core.SecurityContext;
 import javax.ws.rs.core.UriInfo;
 import javax.ws.rs.core.Response.Status;
 
+import org.eclipse.microprofile.jwt.JsonWebToken;
 import org.eclipse.microprofile.openapi.annotations.Operation;
+
 import org.jboss.logging.Logger;
 
 import de.hsos.swa.studiom.StudentsManagement.entity.Student;
@@ -39,15 +44,22 @@ import de.hsos.swa.studiom.shared.exceptions.JoinGroupException;
 
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
-// Matnr wird wegen login überflüssig
-@Path("/api/v1/group/{groupId}/{matNr}")
+
+@Path("/api/v1/group/{groupId}")
 @ApplicationScoped
+
 public class GroupGroupIDRessource {
 
     Logger log = Logger.getLogger(GroupGroupIDRessource.class);
 
     @Context
     UriInfo uriInfo;
+
+    @Context
+    SecurityContext ctx;
+
+    @Inject
+    JsonWebToken jwt;
 
     @Inject
     GroupRepository service;
@@ -57,6 +69,7 @@ public class GroupGroupIDRessource {
 
     @GET
     @Operation(summary = "Find Group with Id")
+    @RolesAllowed("USER")
     public Response getGroup(@PathParam("groupId") int groupId) {
         try {
             log.info("GET " + uriInfo.getPath());
@@ -72,6 +85,7 @@ public class GroupGroupIDRessource {
     }
 
     @POST
+    @RolesAllowed("STUDENT")
     @Operation(summary = "Change a group", description = "Change the name, maxMembers or module of a group")
     public Response changeGroup(@PathParam("groupId") int groupId, NewGroupDTO newGroup) {
         try {
@@ -91,11 +105,17 @@ public class GroupGroupIDRessource {
     }
 
     @POST
+    @RolesAllowed("STUDENT")
     @Path("/student")
     @Operation(summary = "Join a Group", description = "Adds a Student to a group if group is not full yet")
-    public Response addStudent(@PathParam("matNr") int matNr, @PathParam("groupId") int groupId) {
+    public Response addStudent(@PathParam("groupId") int groupId) {
+        log.info("POST " + uriInfo.getPath());
+        Object claim = jwt.getClaim("matNr");
+        if (claim == null) {
+            return Response.status(Status.UNAUTHORIZED).build();
+        }
+        int matNr = Integer.valueOf(claim.toString());
         try {
-            log.info("POST " + uriInfo.getPath());
             Optional<Group> addStudent = service.addStudent(groupId, matNr);
             if (addStudent.isPresent()) {
                 return Response.ok(GroupDTO.Converter.toDTO(addStudent.get())).build();
@@ -109,10 +129,16 @@ public class GroupGroupIDRessource {
     }
 
     @DELETE
+    @RolesAllowed("STUDENT")
     @Operation(summary = "Delete a Group", description = "Deletes a Group, if called by the owner of the group")
-    public Response deleteGroup(@PathParam("groupId") int groupId, @PathParam("matNr") int matNr) {
+    public Response deleteGroup(@PathParam("groupId") int groupId) {
+        log.info("DELETE " + uriInfo.getPath());
+        Object claim = jwt.getClaim("matNr");
+        if (claim == null) {
+            return Response.status(Status.UNAUTHORIZED).build();
+        }
+        int matNr = Integer.valueOf(claim.toString());
         try {
-            log.info("DELETE " + uriInfo.getPath());
             boolean deleted = service.deleteGroup(matNr, groupId);
             if (deleted) {
                 return Response.ok().build();
@@ -125,10 +151,16 @@ public class GroupGroupIDRessource {
 
     @DELETE
     @Path("/student")
+    @RolesAllowed("STUDENT")
     @Operation(summary = "Leave a Group", description = "Leave a Group, if called by a member of the group")
-    public Response removeStudent(@PathParam("matNr") int matNr, @PathParam("groupId") int groupId) {
+    public Response removeStudent(@PathParam("groupId") int groupId) {
+        log.info("DELETE " + uriInfo.getPath());
+        Object claim = jwt.getClaim("matNr");
+        if (claim == null) {
+            return Response.status(Status.UNAUTHORIZED).build();
+        }
+        int matNr = Integer.valueOf(claim.toString());
         try {
-            log.info("DELETE " + uriInfo.getPath());
             Optional<Group> removeStudent = service.removeStudent(groupId, matNr);
             if (removeStudent.isPresent()) {
                 return Response.ok(GroupDTO.Converter.toDTO(removeStudent.get())).build();
